Some improvements
Change the output result: output the aggregate files of all levels at once.
Change the output result: the user can specify the number of aggregates produced by a single execution (using different random seeds).
Improve reliability: Internal tests algorithm for non-overlapping and attached conditions are improved.

+++ aggregate_gen_SA.hpp
@@ -1,5 +1,5 @@
 
 extern std::default_random_engine re;
-extern std::uniform_real_distribution<float> urf;
+extern std::uniform_real_distribution<double> urf;
 
-Eigen::MatrixXf aggregate_gen_SA(const float& a, const int& num_sph, const float& kf, const float& Df, const float& tol);
+Eigen::MatrixXd aggregate_gen_SA(const double& a, const int& num_sph, const double& kf, const double& Df, const double& tol);
